- Defaulting to Parallel, not MPI - Removed the pareto_samples.hpp and moved the content into the nsga2 example (ex_nsga2.cpp)
