+++ examples/ex_nsga2.cpp
@@ -4,7 +4,6 @@
 #include <sferes/ea/nsga2.hpp>
 #include <sferes/eval/eval.hpp>
 #include <sferes/stat/pareto_front.hpp>
-#include <sferes/stat/pareto_samples.hpp>
 #include <sferes/modif/dummy.hpp>
 #include <sferes/run.hpp>
 #include <boost/program_options.hpp>
@@ -13,13 +12,8 @@
 #define NO_PARALLEL
 #endif
 
-#define NO_MPI
-
 #ifndef NO_PARALLEL
 #include <sferes/eval/parallel.hpp>
-#ifndef NO_MPI
-#include <sferes/eval/mpi.hpp>
-#endif
 #else
 #include <sferes/eval/eval.hpp>
 #endif
@@ -75,18 +69,70 @@ public:
   }
 };
 
+namespace sferes {
+    namespace stat {
+        SFERES_STAT(ParetoSamples, Stat)
+        {
+        public:
+            typedef std::vector<boost::shared_ptr<Phen>> pareto_t;
+            // assume a ea.pareto_front() method
+            template <typename E>
+            void refresh(const E& ea)
+            {
+                _pareto_front = ea.pareto_front();
+                parallel::sort(_pareto_front.begin(), _pareto_front.end(),
+                    fit::compare_objs_lex());
+                this->_create_log_file(ea, "pareto_samples.dat");
+                if (ea.dump_enabled())
+                    show_all(*(this->_log_file), ea.gen(), ea.nb_evals());
+                //this->_log_file->close();
+            }
+            void show(std::ostream & os, size_t k) const
+            {
+                os << "log format : gen nb_evals id sample_1 ... sample_n" << std::endl;
+                show_all(os, 0);
+                _pareto_front[k]->develop();
+                _pareto_front[k]->show(os);
+                _pareto_front[k]->fit().set_mode(fit::mode::view);
+                _pareto_front[k]->fit().eval(*_pareto_front[k]);
+                os << "=> displaying individual " << k << std::endl;
+                os << "samples:";
+                for (unsigned j = 0; j < _pareto_front[k]->data().size(); ++j)
+                    os << _pareto_front[k]->data(j) << " ";
+                os << std::endl;
+                assert(k < _pareto_front.size());
+            }
+            const pareto_t& pareto_front() const
+            {
+                return _pareto_front;
+            }
+            template <class Archive>
+            void serialize(Archive & ar, const unsigned int version)
+            {
+                ar& BOOST_SERIALIZATION_NVP(_pareto_front);
+            }
 
+            void show_all(std::ostream & os, size_t gen = 0, size_t nb_evals = 0) const
+            {
+                for (unsigned i = 0; i < _pareto_front.size(); ++i) {
+                    os << gen << " " << nb_evals << " " << i << " ";
+                    for (unsigned j = 0; j < _pareto_front[i]->data().size(); ++j)
+                        os << _pareto_front[i]->data(j) << " ";
+                    os << std::endl;
+                }
+            }
 
+        protected:
+            pareto_t _pareto_front;
+        };
+    }
+}
 
 int main(int argc, char **argv) {
   std::cout<<"running "<<argv[0]<<" ... try --help for options (verbose)"<<std::endl;
 
 #ifndef NO_PARALLEL
-#ifndef NO_MPI
-  typedef eval::Mpi<Params> eval_t;
-#else
   typedef eval::Parallel<Params> eval_t;
-#endif
 #else
   typedef eval::Eval<Params> eval_t;
 #endif
